Extracted className prop into an interface
