--- Servers/UI/OJS.Servers.Ui/ClientApp/src/components/common/IHaveChildrenProps.ts
@@ -1,16 +1,0 @@
-import { ReactNode } from 'react';
-
-interface IHaveChildrenProps {
-    children: ReactNode,
-}
-
-interface IHaveOptionalChildrenProps {
-    children?: ReactNode,
-}
-
-export type {
-    IHaveChildrenProps,
-    IHaveOptionalChildrenProps,
-};
-
-export default IHaveChildrenProps;
