+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/common/Props.ts
@@ -1,0 +1,24 @@
+import { ReactNode } from 'react';
+
+type ClassNameType = string | string[];
+
+interface IHaveChildrenProps {
+    children: ReactNode;
+}
+
+interface IHaveOptionalChildrenProps {
+    children?: ReactNode;
+}
+
+interface IHaveOptionalClassName {
+    className?: ClassNameType;
+}
+
+export type {
+    IHaveChildrenProps,
+    IHaveOptionalChildrenProps,
+    IHaveOptionalClassName,
+    ClassNameType,
+};
+
+// export default IHaveChildrenProps;
