--- Servers/UI/OJS.Servers.Ui/ClientApp/src/components/guidelines/icons/SubmissionSuccessfulIcon.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import { FaCheckCircle } from 'react-icons/fa';
-
-const SubmissionSuccessfulIcon = () => (
-    <FaCheckCircle color="green" />
-);
-
-export default SubmissionSuccessfulIcon;
