+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/hooks/use-loading.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { createContext, useCallback, useContext, useState } from 'react';
-import IHaveChildrenProps from '../components/common/IHaveChildrenProps';
+import { IHaveChildrenProps } from '../components/common/Props';
 import Loading from '../components/guidelines/loading/Loading';
 
 interface ILoadingContext {
